sr/constant moved
