+++ apps/gotask_ui/src/app/(portal)/asset/createAsset/createAsset.tsx
@@ -14,8 +14,8 @@ import { User } from "../../user/interfaces/userInterface";
 import useSWR from "swr";
 import { fetcherUserList } from "../../user/services/userAction";
 import MobileInputs from "./mobileInputs";
-import { ASSET_TYPE, calculateWarrantyDate } from "@/app/common/constants/asset";
-import { systemTypeOptions } from "../assetConstants";
+import { ASSET_TYPE } from "@/app/common/constants/asset";
+import { calculateWarrantyDate, systemTypeOptions } from "../assetConstants";
 import AccessInputs from "./accessInput";
 import PrinterInputs from "./printerInputs";
 import FingerprintScannerInputs from "./fingerPrintInputs";
