+++ apps/gotask_ui/src/app/(portal)/asset/edit/[id]/editAsset.tsx
@@ -15,11 +15,11 @@ import { User } from "@/app/(portal)/user/interfaces/userInterface";
 import useSWR from "swr";
 import { fetcherUserList } from "@/app/(portal)/user/services/userAction";
 import MobileInputs from "../../createAsset/mobileInputs";
-import { ASSET_TYPE, calculateWarrantyDate } from "@/app/common/constants/asset";
+import { ASSET_TYPE } from "@/app/common/constants/asset";
 import HistoryIcon from "@mui/icons-material/History";
 import IssueHistoryDrawer from "../../createIssues/issuesDrawer";
 import { ArrowBack } from "@mui/icons-material";
-import { systemTypeOptions } from "../../assetConstants";
+import { calculateWarrantyDate, systemTypeOptions } from "../../assetConstants";
 import AccessInputs from "../../createAsset/accessInput";
 import PrinterInputs from "../../createAsset/printerInputs";
 import FingerprintScannerInputs from "../../createAsset/fingerPrintInputs";
